Preliminary changes to Idris to support unicode strings
This commit counts as a "add the critical part and break everything" step in
separating out the current String type from a new unicode string type.

I moved the String type and Char type support to CString and CChar types
respectively. CString and CChar are new primitive types. CString represents
essentially the same thing String did but makes it explicit. CChar is a char
type which only represents values that fit into one Bits8. This includes the
ascii range but can be used for octets like in C.

The primitive operations on String, now CString, were essentially unchanged
except for their types and the types of the library functions based on them.
FString now corresponds to CString in the ffi.

The new String type and Char type represent unicode text and unicode characters
respectively. I added very little support operations for these types just in order
to get Show instances to typecheck.

Notable are the changes to the parser and the unification process which allow
character literals and string literals to stand for either Char CChar String or
CString, if it makes sense for that literal. For example 'a' can be Char (the
default) or CChar if Char would not check. But '国' is immediately judged to be
a Char constant by the parser. Using \x-escaped "characters" between 128 and
255 will still allow literals to be recognized as a CChar / CString, and in
CStrings those characters would be one to one with CChar / bytes. Char and
String are the default, so the user should never even know about CString or
CChar unless they are dealing with FFI or lower level OS functions.

New file / stdio actions that work on the new String currently use a utf8encode
or decode function to type check. But in real life this decision needs to be
made by the user's locale.

In the Java code generator the type CChar is mapped to the Java char type, but
this may be inappropriate.

While Idris, the prelude and the several packages now compile, they almost
certainly do not function anymore until the new string operations are
implemented.